fix: erreur de compilation et de type au niveau du serveur, il y avait une tentative de fermeture de socket après avoir libéré la mémoire de cette dernière
- ajout des fichiers produits par make au .gitignore

+++ src/common/request.c
@@ -75,7 +75,7 @@ void prepare_invitation_req(req_invite_group *req, int ID, int ID_group, int gue
         return;
     }
 
-    for (size_t i = 0; i < guest_number; i++)
+    for (int i = 0; i < guest_number; i++)
     {
         req->invited_ids[i] = ((guest_ids[i] & MASK_11_BITS) << 5);
     }
